file utils coverage

+++ create-pfx.sh
@@ -48,9 +48,13 @@ fi
 
 echo -e "${CYAN} [PFX BUILDER SH]: Creating KEY PEM ${NC}"
 openssl pkcs8 -inform der -in $KEY_LOCATION -passin pass:${KEY_PASS} -out key.pem
+[ $? -eq 0 ]  || exit 1
 echo -e "${CYAN} [PFX BUILDER SH]: Creating CER PEM ${NC}"
 openssl x509 -inform der -in $CERT_LOCATION -out cer.pem
+[ $? -eq 0 ]  || exit 1
 openssl pkcs12 -passout pass:${KEY_PASS} -export -in cer.pem -inkey key.pem -out ./keystore/${RFC}.pfx
+[ $? -eq 0 ]  || exit 1
 echo -e "${CYAN} [PFX BUILDER SH]: Removing intermediates ${NC}"
 rm key.pem && rm cer.pem
+[ $? -eq 0 ]  || exit 1
 echo -e "${GREEN} [PFX BUILDER SH]: PFX created at: \"./keystore/${RFC}.pfx${NC}\""
